+++ CHANGELOG.md
@@ -8,6 +8,7 @@ For detailed information check [Blogophon's releases on Github](https://github.c
 * [x] Adding `h-entry` microformats and English template language
 * [x] Added `8<` symbol (for snippets)
 * [X] Codepen.io integration
+* [x] More sharing links (Pocket, Tumblr, Wordpress)
 
 1.1.0
 -----